Refactor default secret types in to internal package
Don't want to make the register function publicly available, ordering issues
might get weird. If people want to use an out-of-tree version, they can
construct it manually.

Make the default cache option in-memory, rather than just dropping things to
align with the signer.
