+++ .gitignore
@@ -1,3 +1,4 @@
 .gobincache
 /bin/golangci-lint*
 /*.test
+go.work*
