+++ internal/platformsecrets/cache_darwin_cgo.go
@@ -1,6 +1,6 @@
 //go:build darwin && cgo
 
-package clitoken
+package platformsecrets
 
 import (
 	"bytes"
