+++ internal/platformsecrets/mem_cache_test.go
@@ -1,0 +1,15 @@
+package platformsecrets
+
+import (
+	"testing"
+)
+
+func TestMemoryWriteThroughCredentialCache(t *testing.T) {
+	cache := &MemCredentialCache{}
+
+	if !cache.Available() {
+		t.Fatal("cache is not available")
+	}
+
+	TestCache(t, cache)
+}
