+++ internal/platformsecrets/signer_darwin.go
@@ -1,6 +1,6 @@
 //go:build darwin && cgo
 
-package clitoken
+package platformsecrets
 
 import (
 	"crypto"
