+++ keychainsecrets/auto/auto.go
@@ -1,0 +1,1 @@
+package auto
