+++ keychainsecrets/go.mod
@@ -6,13 +6,14 @@ replace lds.li/oauth2ext => ../
 
 require (
 	golang.org/x/oauth2 v0.34.0
+	lds.li/keychain v0.0.0-20260125191015-a9e54317a85c
 	lds.li/oauth2ext v0.0.0-00010101000000-000000000000
 )
 
 require (
 	github.com/ebitengine/purego v0.9.1 // indirect
+	github.com/google/go-cmp v0.7.0 // indirect
 	github.com/tink-crypto/tink-go/v2 v2.6.0 // indirect
 	golang.org/x/crypto v0.46.0 // indirect
 	google.golang.org/protobuf v1.36.11 // indirect
-	lds.li/keychain v0.0.0-20260125161459-7523d5fe457b // indirect
 )
